Tonnes of new features added - read CHANGES.rst file

+++ cheermonk/blueprints/user/models.py
@@ -1,29 +1,89 @@
+import datetime
+from collections import OrderedDict
 from hashlib import md5
+
+import pytz
 from flask import current_app
+
 from flask_login import UserMixin
+
 from itsdangerous import URLSafeTimedSerializer, \
     TimedJSONWebSignatureSerializer
 
-from cheermonk.lib.util_sqlalchemy import ResourceMixin
+from sqlalchemy import or_
+
+from cheermonk.lib.util_sqlalchemy import ResourceMixin, AwareDateTime
+from cheermonk.blueprints.billing.models.credit_card import CreditCard
+from cheermonk.blueprints.billing.models.subscription import Subscription
+from cheermonk.blueprints.billing.models.invoice import Invoice
 from cheermonk.extensions import db, bcrypt
 
 
 class User(UserMixin, ResourceMixin, db.Model):
+    ROLE = OrderedDict([
+        ('guest', 'Guest'),
+        ('member', 'Member'),
+        ('admin', 'Admin')
+    ])
 
     __tablename__ = 'users'
     id = db.Column(db.Integer, primary_key=True)
 
+    # Relationships.
+    credit_card = db.relationship(CreditCard, uselist=False, backref='users',
+                                  passive_deletes=True)
+    subscription = db.relationship(Subscription, uselist=False,
+                                   backref='users', passive_deletes=True)
+    invoices = db.relationship(Invoice, backref='users', passive_deletes=True)
+
     # Authentication.
+    role = db.Column(db.Enum(*ROLE, name='role_types'),
+                     index=True, nullable=False, server_default='member')
+    active = db.Column('is_active', db.Boolean(), nullable=False,
+                       server_default='1')
     username = db.Column(db.String(24), unique=True, index=True)
     email = db.Column(db.String(255), unique=True, index=True, nullable=False,
                       server_default='')
     password = db.Column(db.String(128), nullable=False, server_default='')
+
+    # Billing.
+    name = db.Column(db.String(128), index=True)
+    payment_id = db.Column(db.String(128), index=True)
+    cancelled_subscription_on = db.Column(AwareDateTime())
+
+    # Activity tracking.
+    sign_in_count = db.Column(db.Integer, nullable=False, default=0)
+    current_sign_in_on = db.Column(AwareDateTime())
+    current_sign_in_ip = db.Column(db.String(45))
+    last_sign_in_on = db.Column(AwareDateTime())
+    last_sign_in_ip = db.Column(db.String(45))
+
+    # Locale.
+    locale = db.Column(db.String(5), nullable=False, server_default='en')
 
     def __init__(self, **kwargs):
         # Call Flask-SQLAlchemy's constructor.
         super(User, self).__init__(**kwargs)
 
         self.password = User.encrypt_password(kwargs.get('password', ''))
+
+    @classmethod
+    def search(cls, query):
+        """
+        Search a resource by 1 or more fields.
+
+        :param query: Search query
+        :type query: str
+        :return: SQLAlchemy filter
+        """
+        if not query:
+            return ''
+
+        search_query = '%{0}%'.format(query)
+        search_chain = (User.email.ilike(search_query),
+                        User.name.ilike(search_query))
+
+        return or_(*search_chain)
 
     @classmethod
     def find_by_identity(cls, identity):
@@ -70,10 +130,35 @@ class User(UserMixin, ResourceMixin, db.Model):
             return None
 
     @classmethod
+    def is_last_admin(cls, user, new_role, new_active):
+        """
+        Determine whether or not this user is the last admin account.
+
+        :param user: User being tested
+        :type user: User
+        :param new_role: New role being set
+        :type new_role: str
+        :param new_active: New active status being set
+        :type new_active: bool
+        :return: bool
+        """
+        is_changing_roles = user.role == 'admin' and new_role != 'admin'
+        is_changing_active = user.active is True and new_active is None
+
+        if is_changing_roles or is_changing_active:
+            admin_count = User.query.filter(User.role == 'admin').count()
+            active_count = User.query.filter(User.is_active is True).count()
+
+            if admin_count == 1 or active_count == 1:
+                return True
+
+        return False
+
+    @classmethod
     def bulk_delete(cls, ids):
         """
         Override the general bulk_delete method because we need to delete them
-        one at a time.
+        one at a time while also deleting them on Stripe.
 
         :param ids: List of ids to be deleted
         :type ids: list
@@ -87,7 +172,16 @@ class User(UserMixin, ResourceMixin, db.Model):
             if user is None:
                 continue
 
-            user.delete()
+            if user.payment_id is None:
+                user.delete()
+            else:
+                subscription = Subscription()
+                cancelled = subscription.cancel(user=user)
+
+                # If successful, delete it locally.
+                if cancelled:
+                    user.delete()
+
             delete_count += 1
 
         return delete_count
@@ -104,12 +198,21 @@ class User(UserMixin, ResourceMixin, db.Model):
         u = User.find_by_identity(identity)
         reset_token = u.serialize_token()
 
-        # This prevents circular imports
+        # This prevents circular imports.
         from cheermonk.blueprints.user.tasks import deliver_password_reset_email
 
         deliver_password_reset_email.delay(u.id, reset_token)
 
         return u
+
+    def is_active(self):
+        """
+        Return whether or not the user account is active, this satisfies
+        Flask-Login by overwriting the default value.
+
+        :return: bool
+        """
+        return self.active
 
     def get_auth_token(self):
         """
@@ -157,3 +260,22 @@ class User(UserMixin, ResourceMixin, db.Model):
 
         serializer = TimedJSONWebSignatureSerializer(private_key, expiration)
         return serializer.dumps({'user_email': self.email}).decode('utf-8')
+
+    def update_activity_tracking(self, ip_address):
+        """
+        Update various fields on the user that's related to meta data on his
+        account, such as the sign in count and ip address, etc..
+
+        :param ip_address: IP address
+        :type ip_address: str
+        :return: SQLAlchemy commit results
+        """
+        self.sign_in_count += 1
+
+        self.last_sign_in_on = self.current_sign_in_on
+        self.last_sign_in_ip = self.current_sign_in_ip
+
+        self.current_sign_in_on = datetime.datetime.now(pytz.utc)
+        self.current_sign_in_ip = ip_address
+
+        return self.save()
